support brownfield apps: implement RestartManager, code refactoring

+++ android/app/src/main/java/com/microsoft/codepush/react/CodePushUpdateResponseUpdateInfo.java
@@ -25,7 +25,7 @@ public class CodePushUpdateResponseUpdateInfo {
     public String Label;
 
     @SerializedName("packageSize")
-    public int PackageSize;
+    public long PackageSize;
 
     @SerializedName("updateAppVersion")
     public boolean UpdateAppVersion;
